+++ games_services/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update minimum support for iOS and Android.
 - Fix the example.
 - Add loadPreviousOccurrence method for iOS to retrieve the previous occurrence of a player's score from a leaderboard.
+- Add ignoreImages to loadAchievements, offloading blocking loading images to IO threads pool.
 
 ## 4.1.1